+++ test/src/customizer/avatar_maker_save_widget_test.mocks.dart
@@ -369,6 +369,15 @@ class MockAvatarMakerController extends _i1.Mock
       ) as String);
 
   @override
+  void randomizedSelectedOptions() => super.noSuchMethod(
+        Invocation.method(
+          #randomizedSelectedOptions,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
   void update([
     List<Object>? ids,
     bool? condition = true,